+++ .gitignore
@@ -1,6 +1,6 @@
 # Default ignored files
 /shelf/
-/workspace.xml
+/.idea/workspace.xml
 # User-specific stuff
 .idea/**/workspace.xml
 .idea/**/tasks.xml
