+++ src/dto/Funding.java
@@ -1,4 +1,4 @@
-package DTO;
+package dto;
 
 import java.time.*;
