+++ src/dto/User.java
@@ -1,4 +1,4 @@
-package DTO;
+package dto;
 
 public class User {
 	private String id;
