+++ src/exceptions/IdAlreadyExistsException.java
@@ -1,4 +1,4 @@
-package Exception;
+package exceptions;
 
 public class IdAlreadyExistsException extends Exception{
     public IdAlreadyExistsException(){
