+++ src/exceptions/InvalidAmountException.java
@@ -1,4 +1,4 @@
-package Exception;
+package exceptions;
 
 public class InvalidAmountException extends Exception{
     public InvalidAmountException(){
