+++ src/exceptions/NotEnoughMoneyException.java
@@ -1,4 +1,4 @@
-package Exception;
+package exceptions;
 
 public class NotEnoughMoneyException extends Exception{
     public NotEnoughMoneyException(){
