+++ src/exceptions/WrongPasswordException.java
@@ -1,4 +1,4 @@
-package Exception;
+package exceptions;
 
 public class WrongPasswordException extends Exception{
     public WrongPasswordException(){
