+++ src/util/UserInput.java
@@ -1,4 +1,4 @@
-package UTIL;
+package util;
 
 import java.util.Scanner;
